FIX 2269 disable the breadcrumb when already on the breadcrumb's view (#2280)
* FIX-2269 disable onClick when using breadcrum on current view

* FIX-2269 move breadcrumbs into own file

* FIX-2269 add log curve values view as breadcrumb
